decameal/users/api/views/login.py
from django.contrib.auth import authenticate
from django.core.exceptions import ObjectDoesNotExist
from rest_framework import generics, status
from rest_framework.authtoken.models import Token
from rest_framework.permissions import AllowAny
from rest_framework.response import Response

from ...models import User
from ..serializers.login_serializer import LoginSerializer


class LoginView(generics.GenericAPIView):
    serializer_class = LoginSerializer
    permission_classes = (AllowAny,)

    def post(self, request):
        email = request.data.get("email", "")
        password = request.data.get("password", "")
        serializer = self.serializer_class(data=request.data)
        if not serializer.is_valid():
            return Response(
                {"message": "failure", "data": "null", "errors": serializer.errors},
                status=status.HTTP_400_BAD_REQUEST,
            )
        try:
            user_object = User.objects.get(email=email)
        except ObjectDoesNotExist:
            return Response(
                {
                    "message": "failure",
                    "data": "null",
                    "error": {"email": "Email does not exist"},
                },
                status=status.HTTP_400_BAD_REQUEST,
            )
        user = authenticate(username=user_object.username, password=password)
        if not user:
            return Response(
                {
                    "message": "failure",
                    "data": "null",
                    "error": {"password": "Incorrect password"},
                },
                status=status.HTTP_400_BAD_REQUEST,
            )
        if not user.is_verified:
            return Response(
                {
                    "message": "failure",
                    "data": "null",
                    "error": {"email": "Email is not yet verified"},
                },
                status=status.HTTP_400_BAD_REQUEST,
            )
        token, _ = Token.objects.get_or_create(user=user)
        return Response(
            {
                "message": "success",
                "data": {"token": token.key},
                "errors": "null",
            },
            status=status.HTTP_200_OK,
        )
